Fix case where we'd get double deletes in unfavorited drawer
