+++ packages/mobile/src/store/offline-downloads/sagas/rehydrateOfflineDataSaga.ts
@@ -11,7 +11,6 @@ import { call, select, put } from 'typed-redux-saga'
 import {
   DOWNLOAD_REASON_FAVORITES,
   getCollectionJson,
-  getLocalCollectionJsonPath,
   getTrackJson
 } from 'app/services/offline-downloader'
